refactor: extract session/message domain to repository/session.rs
